Fix bug when adding rooms
Previously adding more than one room in a session caused multiple empty
rooms to be created.

The problem was a result of how the submit event was intercepted. It was
installed EVERY SINGLE TIME the modal was displayed, meaning that by
the fourth room, it would submit it four times, often with null data.

This commit address this issue by uninstalling the event handler after
it fires.